+++ src/index-globals.js
@@ -1,6 +1,16 @@
 'use strict';
 
+const path = require('path');
+const processCwd = process.cwd();
+
 const globals = {
+  sourceFiles: [
+    path.join(processCwd, 'bin/**/*.js'),
+    path.join(processCwd, 'src/**/*.js'),
+    path.join(processCwd, 'gulp/**/*.js'),
+    path.join(processCwd, 'gulpfile.js')
+  ],
+
   babel: {
     presets: [
       require.resolve('babel-preset-es2015'),
@@ -20,12 +30,4 @@ if (process.env.urbanJSToolGlobals) {
   Object.assign(globals, JSON.parse(process.env.urbanJSToolGlobals));
 }
 
-module.exports = {
-  get babel() {
-    return globals.babel;
-  },
-
-  set babel(value) {
-    globals.babel = value;
-  }
-};
+module.exports = globals;